src/requirements.txt
feedparser==5.2.1
PyYAML==3.12
